requirements.txt
Flask==2.3.3
gunicorn==21.2.0
PyPDF2==3.0.0
pdf2docx==0.5.6
Pillow==10.1.0
python-docx==1.1.0
python-dotenv==1.0.0
